backend/requirements.txt
flask
sqlalchemy
psycopg2-binary
twilio
python-dotenv
flask-cors
google-cloud-aiplatform
